+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### ✨ Improved
 
 * [#10](https://github.com/sdss/lvmgort/pull/10) `GortObserver.slew()` and the standard iteration task now calculate the adjusted target coordinates for the spec telescope so that the slew puts the star on top of the desired fibre to be observed. Most of this relies on code copied from `lvmtipo` used to calculate the field angle through the siderostat.
+* `SkyCoordinates` now accepts a `name` attribute that is passed to the `SKYENAME` and `SKYWNAME` header keywords.
 
 ### 🔧 Fixed